add reports for each browser
